+++ nsone/rest/data.py
@@ -5,21 +5,6 @@
 #
 from . import resource
 
-class Job(resource.BaseResource):
-
-    ROOT = 'monitoring/jobs'
-    PASSTHRU_FIELDS = ['name', 'config']
-
-    def list(self, callback=None, errback=None):
-        return self._make_request('GET', '%s' % (self.ROOT),
-                                  callback=callback,
-                                  errback=errback)
-
-    def retrieve(self, jobid, callback=None, errback=None):
-        return self._make_request('GET', '%s/%s' % (self.ROOT, jobid),
-                                  callback=callback,
-                                  errback=errback)
-
 class Source(resource.BaseResource):
 
     ROOT = 'data/sources'
